+++ color/color.hpp
@@ -12,9 +12,9 @@ struct Color {
     /**
      * @brief Get the Luminance value
      * 
-     * @return float 
+     * @return double 
      */
-    virtual float getLuminance() const = 0;
+    virtual double getLuminance() const = 0;
 
     /**
      * @brief Convert Color to String
@@ -29,9 +29,9 @@ struct Color {
 ostream& operator<<(ostream& os, const Color& c);
 
 struct RGB : public Color {
-    float red;
-    float green;
-    float blue;
+    double red;
+    double green;
+    double blue;
     
     /**
      * @brief Construct a new RGB object
@@ -40,7 +40,7 @@ struct RGB : public Color {
      * @param _green 
      * @param _blue 
      */
-    inline RGB(float _red, float _green, float _blue) : red(_red), green(_green), blue(_blue) {}
+    inline RGB(double _red, double _green, double _blue) : red(_red), green(_green), blue(_blue) {}
 
     /**
      * @brief Construct a new RGB object
@@ -48,7 +48,7 @@ struct RGB : public Color {
      */
     inline RGB() : red(0), green(0), blue(0) {}
 
-    inline float getLuminance() const override { return 0.27 * red + 0.67 * green + 0.06 * blue; } 
+    inline double getLuminance() const override { return 0.27 * red + 0.67 * green + 0.06 * blue; } 
 
     string toString() const override {
         ostringstream ss;
@@ -59,8 +59,8 @@ struct RGB : public Color {
     }
 };
 
-inline RGB clamp(const RGB& c, float v) { return RGB(clampHigh(c.red, v), clampHigh(c.green, v), clampHigh(c.blue, v)); };
-inline RGB equalize(const RGB& c, float v) { return RGB(c.red / v, c.green / v, c.blue / v); };
-RGB equalizeClamp(const RGB& c, float v);
-RGB gammaCorrection(const RGB& c, float k, float gamma);
-RGB gammaClamp(const RGB& c, float k, float gamma);
+inline RGB clamp(const RGB& c, double v) { return RGB(clampHigh(c.red, v), clampHigh(c.green, v), clampHigh(c.blue, v)); };
+inline RGB equalize(const RGB& c, double v) { return RGB(c.red / v, c.green / v, c.blue / v); };
+RGB equalizeClamp(const RGB& c, double v);
+RGB gammaCorrection(const RGB& c, double k, double gamma);
+RGB gammaClamp(const RGB& c, double k, double gamma);
